src/python/requirements.txt
grpcio==1.56.0
protobuf==4.23.4

